Deduplicate HTTP fetch and validation utilities
- Extract shared fetchUrl, isHtmlContent, and timeout constants to
  src/server/http/fetch.ts
- Extract shared escapeHtml utility to src/server/http/html.ts
- Extract shared feedUrlSchema to src/server/trpc/validation.ts
- Fix inconsistent User-Agent in saved.ts (was hardcoded, now uses
  shared USER_AGENT constant)
- Remove unused useSwipeGestures hook (never imported anywhere)

Net reduction of ~224 lines of duplicated code.
